Parse map and vector literals, handle knock-on effects
Before now, collection "literals" were really just elements on the stack
that got consumed by special builtins ([] and {}) to construct the
objects. This was brittle and didn't allow for nesting. The fix: just
pad the tokens the way it was already being done for parentheses.

Having a way to go from stack to collection object is nice in certain
situations, so [] and {} have been renamed rather than removed.