Make `cond` take a vector of arms now that that's possible
Before it was possible to have a vector literal with functions as
elements, the second-cleanest way to do multi-armed branching was to lay
the arms on the stack and explicitly specify how many there were.

We *could* take a vector of vectors of functions, but special-casing the
"else" branch (which gives the vector odd length) isn't too bad.