Completely(?) unify builtins!
By treating functions as their sequences of instructions rather than
tokens, they can be freely curried and composed without becoming opaque.
Builtins are still Crystal code, but it's very easy to pretend otherwise
by converting them to a single-instruction Phelix instance that invokes
said builtin.

Inspection-wise, this is really nice; debug-printing any function (whether
builtin, prelude, or user-defined) will show what that function would do if
called. The "downside" is that it becomes pretty much impossible to tell
the provenance of a function without invoking the `source` builtin, but
that was more or less the goal all along.