Make `source` way more useful on non-builtins
For builtins, the obvious thing for `source` to do is return the Crystal
code that implements them. For all other functions, though, it'd be nice
if it returned a vector of instructions that could be manipulated and
re-evaluated as Phelix code. This patch very nearly gets there, but we
still need some sort of symbol type to differentiate between "mention"
and invoke, as well as a way to convert from vector back to function.