Add dispatch mechanism for overloadable builtins
There are a handful of cases where it's nice to define multiple
variations of a word and have "the right one" fire based on the types of
the values at the top of the stack; being able to invoke `nth` whether
the stack is `[... Vec Num]` or `[... Num Vec]` was the driving example.

Most builtins *won't* be overloaded, so care has been taken to avoid
incurring the performance penalty of dispatching where unnecessary.